+++ addons/ace_interact/CfgVehicles.hpp
@@ -12,19 +12,5 @@ class CfgVehicles {
                 icon = "\a3\Ui_f\data\GUI\Cfg\CommunicationMenu\call_ca.paa";
             };
         };
-
-        class ACE_Actions {
-            class ACE_MainActions {
-                class ACRE_InfantryPhone {
-                    displayName = CSTRING(infantryPhone);
-                    condition = QUOTE(_player call FUNC(checkIntercomInfantry));
-                    exceptions[] = {"isNotInside", "isNotSitting"};
-                    statement = "true";
-                    insertChildren = QUOTE(_this call FUNC(infantryPhoneChildrenActions));
-                    priority = 0.1;
-                    icon = "\a3\Ui_f\data\GUI\Cfg\CommunicationMenu\call_ca.paa";
-                 };
-             };
-         };
     };
 };
