+++ addons/ace_interact/XEH_PREP.hpp
@@ -1,8 +1,5 @@
 PREP(actionSetMTT);
 PREP(actionSetSpatialAudio);
-PREP(checkIntercomInfantry);
-PREP(infantryPhoneAction);
-PREP(infantryPhoneChildrenActions);
 PREP(generateSpatialChildrenActions);
 PREP(radioChildrenActions);
 PREP(radioListChildrenActions);
