+++ addons/ace_interact/stringtable.xml
@@ -141,25 +141,5 @@
             <French>Définir comme Multi PTT</French>
             <Korean>%1번 PTT로 설정</Korean>
         </Key>
-        <Key ID="STR_ACRE_ace_interact_infantryPhone">
-            <English>Infantry Phone</English>
-            <Spanish>Teléfono de infanteria</Spanish>
-            <German>Außenbord Sprechanlage</German>
-        </Key>
-        <Key ID="STR_ACRE_ace_interact_giveInfantryPhone">
-            <English>Pass Infantry Phone</English>
-            <Spanish>Dar teléfono de infanteria</Spanish>
-            <German>Außenbord Sprechanlage geben</German>
-        </Key>
-        <Key ID="STR_ACRE_ace_interact_takeInfantryPhone">
-            <English>Take Infantry Phone</English>
-            <Spanish>Coger teléfono de infanteria</Spanish>
-            <German>Außenbord Sprechanlage abnehmen</German>
-        </Key>
-        <Key ID="STR_ACRE_ace_interact_returnInfantryPhone">
-            <English>Return Infantry Phone</English>
-            <Spanish>Colgar teléfono de infanteria</Spanish>
-            <German>Außenbord Sprechanlage aufhängen</German>
-        </Key>
     </Package>
 </Project>
