+++ addons/sys_attenuate/CfgVehicles.hpp
@@ -30,11 +30,9 @@ class CfgVehicles {
                     Compartment4 = 0;
                 };
             };
-            class CVC {
-                hasCVC = false;
-            };
         };
     };
+
     class Car_F: Car {};
     class Wheeled_APC_F: Car_F {
         class ACRE {
@@ -63,9 +61,6 @@ class CfgVehicles {
                     Compartment3 = 1;
                     Compartment4 = 1;
                 };
-            };
-            class CVC {
-                hasCVC = true;
             };
         };
     };
@@ -98,82 +93,6 @@ class CfgVehicles {
                     Compartment4 = 1;
                 };
             };
-            class CVC {
-                hasCVC = true;
-                hasInfantryPhone = true;
-            };
-        };
-    };
-    class Tank_F: Tank {};
-    class MBT_01_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.35, -4.4, -1};
-            };
-        };
-    };
-    class MBT_01_arty_base_F: MBT_01_base_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.35, -4.86, -1.4};
-            };
-        };
-    };
-    class MBT_01_mlrs_base_F: MBT_01_base_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.35, -4.43, -0.33};
-            };
-        };
-    };
-    class APC_Tracked_01_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {-1.1, -4.86, -0.82};
-            };
-        };
-    };
-    class B_APC_Tracked_01_base_F: APC_Tracked_01_base_F {};
-    class B_APC_Tracked_01_AA_F: B_APC_Tracked_01_base_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {-1.1, -4.85, -1.14};
-            };
-        };
-    };
-    class MBT_02_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.38, -4.77, -1.1};
-            };
-        };
-    };
-    class MBT_02_arty_base_F: MBT_02_base_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.4, -5.4, -1.65};
-            };
-        };
-    };
-    class APC_Tracked_02_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {0.98, -4.9, -0.79};
-            };
-        };
-    };
-    class MBT_03_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.53, -5.67, -1.29};
-            };
-        };
-    };
-    class APC_Tracked_03_base_F: Tank_F {
-        class ACRE: ACRE {
-            class CVC: CVC {
-                infantryPhonePosition[] = {1.1, -3.87, -0.78};
-            };
         };
     };
 
@@ -204,9 +123,6 @@ class CfgVehicles {
                     Compartment3 = 0;
                     Compartment4 = 0;
                 };
-            };
-            class CVC {
-                hasCVC = true;
             };
         };
     };
@@ -239,9 +155,6 @@ class CfgVehicles {
                     Compartment4 = 0;
                 };
             };
-            class CVC {
-                hasCVC = true;
-            };
         };
     };
 
@@ -272,10 +185,6 @@ class CfgVehicles {
                     Compartment3 = 0;
                     Compartment4 = 0;
                 };
-            };
-            class CVC {
-                hasCVC = true;
-                hasInfantryPhone = true;
             };
         };
     };
@@ -308,9 +217,6 @@ class CfgVehicles {
                     Compartment4 = 0;
                 };
             };
-            class CVC {
-                hasCVC = false;
-            };
         };
     };
 
@@ -342,9 +248,6 @@ class CfgVehicles {
                     Compartment3 = 0;
                     Compartment4 = 0;
                 };
-            };
-            class CVC {
-                hasCVC = true;
             };
         };
     };
