+++ addons/sys_intercom/$PBOPREFIX$
@@ -1,0 +1,1 @@
+idi\acre\addons\sys_intercom
