+++ addons/sys_intercom/XEH_PREP.hpp
@@ -1,0 +1,4 @@
+PREP(checkInfantryPhoneAvailability);
+PREP(infantryPhoneAction);
+PREP(infantryPhoneSoundCall);
+PREP(infantryPhoneChildrenActions);
